Updated index.html and style.css
Removed project-tile divs and gave a elements in project-content the project-tile class. changed text in javascript-projects and frontEnd_libraries_projects sections from p to a elements.

Scrapped styling for project tiles and moved to selector for a elements, added tiles for badge elements, tweaked heights of sections and content justification/alignment.